# backend/openapi.yaml
openapi: 3.0.3
info:
  title: FlexLiving Reviews API
  description: |
    A comprehensive review management system for FlexLiving properties with integrations for Hostaway and Google Reviews.
    
    This API provides endpoints for managing property reviews, statistics, and external integrations.
    
    ## Features
    - Review management with filtering and pagination
    - Property management with review aggregation
    - Hostaway integration for review synchronization
    - Google Places API integration for review exploration
    - Comprehensive statistics and analytics
    
  version: 1.0.0
  contact:
    name: FlexLiving Development Team
    email: dev@flexliving.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001
    description: Development server
  - url: ""
    description: Production server (to be configured)
...
paths:
...
  /api/reviews:
    get:
      tags:
        - Reviews
      summary: Get all reviews with filtering and pagination
      description: Retrieve reviews with optional filtering by property, channel, rating, status, and date range
      operationId: getAllReviews
      parameters:
        - name: listingId
          in: query
          description: Filter by property listing ID
          required: false
          schema:
            type: string
            example: "PROP001"
        - name: channel
          in: query
          description: Filter by booking channel
          required: false
          schema:
            type: string
            enum: [Airbnb, Booking.com, Direct, Vrbo, Expedia, Google, Other]
            example: "Airbnb"
        - name: rating
          in: query
          description: Filter by minimum rating
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10
            example: 8
        - name: status
          in: query
          description: Filter by review status
          required: false
          schema:
            type: string
            enum: [published, pending, rejected]
            example: "published"
        - name: showOnWebsite
          in: query
          description: Filter by website visibility
          required: false
          schema:
            type: boolean
            example: true
        - name: startDate
          in: query
          description: Filter reviews from this date (ISO 8601)
          required: false
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - name: endDate
          in: query
          description: Filter reviews until this date (ISO 8601)
          required: false
          schema:
            type: string
            format: date
            example: "2024-12-31"
        - name: sortBy
          in: query
          description: Sort field
          required: false
          schema:
            type: string
            enum: [submittedAt, rating, guestName, channel]
            default: submittedAt
            example: "submittedAt"
        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
            example: "desc"
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReviewsResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/reviews/{id}:
    get:
      tags:
        - Reviews
      summary: Get review by ID
      description: Retrieve a specific review by its ID
      operationId: getReviewById
      parameters:
        - name: id
          in: path
          description: Review ID
          required: true
          schema:
            type: string
            example: "60f7b3b3b3b3b3b3b3b3b3b3"
      responses:
        '200':
          description: Review retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSuccessResponse'
        '400':
          description: Invalid review ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Reviews
      summary: Update review
      description: Update review properties such as status, website visibility, and response text
      operationId: updateReview
      parameters:
        - name: id
          in: path
          description: Review ID
          required: true
          schema:
            type: string
            example: "60f7b3b3b3b3b3b3b3b3b3b3"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
            example:
              showOnWebsite: true
              responseText: "Thank you for your feedback!"
              status: "published"
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSuccessResponse'
        '400':
          description: Invalid request data or review ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/reviews/{id}/helpful:
    post:
      tags:
        - Reviews
      summary: Mark review as helpful or not helpful
      description: Update the helpful/not helpful count for a review
      operationId: markReviewHelpful
      parameters:
        - name: id
          in: path
          description: Review ID
          required: true
          schema:
            type: string
            example: "60f7b3b3b3b3b3b3b3b3b3b3"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelpfulRequest'
            example:
              helpful: true
      responses:
        '200':
          description: Review helpful status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSuccessResponse'
        '400':
          description: Invalid request data or review ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/properties:
    get:
      tags:
        - Properties
      summary: Get all properties
      description: Retrieve all properties with optional filtering and sorting
      operationId: getAllProperties
      parameters:
        - name: isActive
          in: query
          description: Filter by active status
          required: false
          schema:
            type: boolean
            example: true
        - name: sortBy
          in: query
          description: Sort field
          required: false
          schema:
            type: string
            enum: [name, avgRating, totalReviews, createdAt]
            default: avgRating
            example: "avgRating"
        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
            example: "desc"
      responses:
        '200':
          description: Properties retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertiesSuccessResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/properties/{id}:
    get:
      tags:
        - Properties
      summary: Get property by ID with reviews
      description: Retrieve a specific property by its ID along with associated reviews
      operationId: getPropertyById
      parameters:
        - name: id
          in: path
          description: Property ID
          required: true
          schema:
            type: string
            example: "60f7b3b3b3b3b3b3b3b3b3b3"
      responses:
        '200':
          description: Property retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyWithReviewsSuccessResponse'
        '400':
          description: Invalid property ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Property not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Properties
      summary: Update property
      description: Update property information
      operationId: updateProperty
      parameters:
        - name: id
          in: path
          description: Property ID
          required: true
          schema:
            type: string
            example: "60f7b3b3b3b3b3b3b3b3b3b3"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePropertyRequest'
            example:
              name: "Updated Modern Downtown Apartment"
              description: "Newly renovated apartment in the heart of downtown"
              isActive: true
      responses:
        '200':
          description: Property updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertySuccessResponse'
        '400':
          description: Invalid request data or property ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Property not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          example: "success"
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data (varies by endpoint)
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: "error"
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          description: Additional error details
          example: "Validation failed"
    ValidationErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: "error"
        message:
          type: string
          example: "Validation failed"
        error:
          type: string
          description: JSON string containing validation errors
          example: '[{"field":"rating","message":"Rating must be between 1 and 10","value":15}]'

    # Health Response
...
    ReviewCategory:
      type: object
      properties:
        category:
          type: string
          enum: [cleanliness, accuracy, check_in, communication, location, value, respect_house_rules]
          example: "cleanliness"
        rating:
          type: integer
          minimum: 1
          maximum: 10
          example: 9
    Review:
      type: object
      properties:
        _id:
          type: string
          example: "60f7b3b3b3b3b3b3b3b3b3b3"
        externalId:
          type: integer
          example: 12345
        type:
          type: string
          enum: [guest-to-host, host-to-guest]
          example: "guest-to-host"
        status:
          type: string
          enum: [published, pending, rejected]
          example: "published"
        rating:
          type: integer
          minimum: 1
          maximum: 10
          example: 9
        publicReview:
          type: string
          example: "Great place to stay! Very clean and comfortable."
        privateReview:
          type: string
          nullable: true
          example: "Minor issue with WiFi speed"
        reviewCategory:
          type: array
          items:
            $ref: '#/components/schemas/ReviewCategory'
        submittedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        guestName:
          type: string
          example: "John Doe"
        listingId:
          type: string
          example: "PROP001"
        listingName:
          type: string
          example: "Modern Downtown Apartment"
        channel:
          type: string
          enum: [Airbnb, Booking.com, Direct, Vrbo, Expedia, Google, Other]
          example: "Airbnb"
        reservationId:
          type: string
          example: "RES123456"
        showOnWebsite:
          type: boolean
          example: true
        responseText:
          type: string
          nullable: true
          example: "Thank you for your feedback!"
        respondedAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-16T09:00:00.000Z"
        helpful:
          type: integer
          minimum: 0
          example: 5
        notHelpful:
          type: integer
          minimum: 0
          example: 1
        source:
          type: string
          example: "hostaway"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"

    # Request Types
    UpdateReviewRequest:
      type: object
      properties:
        showOnWebsite:
          type: boolean
          example: true
        responseText:
          type: string
          maxLength: 1000
          example: "Thank you for your feedback!"
        status:
          type: string
          enum: [published, pending, rejected]
          example: "published"
    HelpfulRequest:
      type: object
      required:
        - helpful
      properties:
        helpful:
          type: boolean
          description: True to mark as helpful, false to mark as not helpful
          example: true

    # Response Types
    ReviewSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Review'
    PaginationParams:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 10
        total:
          type: integer
          minimum: 0
          example: 150
        pages:
          type: integer
          minimum: 0
          example: 15
    PaginatedReviewsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            pagination:
              $ref: '#/components/schemas/PaginationParams'

    # Statistics Types
...
    Property:
      type: object
      properties:
        _id:
          type: string
          example: "60f7b3b3b3b3b3b3b3b3b3b3"
        externalId:
          type: string
          example: "PROP001"
        name:
          type: string
          example: "Modern Downtown Apartment"
        address:
          type: string
          example: "123 Main St, Downtown, City 12345"
        type:
          type: string
          enum: [Apartment, Studio, House, Penthouse, Villa, Other]
          example: "Apartment"
        bedrooms:
          type: integer
          minimum: 0
          maximum: 20
          example: 2
        bathrooms:
          type: integer
          minimum: 0
          maximum: 20
          example: 2
        maxGuests:
          type: integer
          minimum: 1
          maximum: 50
          example: 4
        imageUrl:
          type: string
          nullable: true
          example: "https://example.com/property-image.jpg"
        description:
          type: string
          nullable: true
          example: "Beautiful modern apartment in the heart of downtown"
        amenities:
          type: array
          items:
            type: string
          example: ["WiFi", "Kitchen", "Parking", "Air Conditioning"]
        avgRating:
          type: number
          format: float
          nullable: true
          example: 8.5
        totalReviews:
          type: integer
          nullable: true
          example: 25
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    UpdatePropertyRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: "Updated Modern Downtown Apartment"
        address:
          type: string
          minLength: 1
          maxLength: 500
          example: "123 Updated Main St, Downtown, City 12345"
        type:
          type: string
          enum: [Apartment, Studio, House, Penthouse, Villa, Other]
          example: "Apartment"
        bedrooms:
          type: integer
          minimum: 0
          maximum: 20
          example: 3
        bathrooms:
          type: integer
          minimum: 0
          maximum: 20
          example: 2
        maxGuests:
          type: integer
          minimum: 1
          maximum: 50
          example: 6
        imageUrl:
          type: string
          maxLength: 1000
          example: "https://example.com/updated-property-image.jpg"
        description:
          type: string
          maxLength: 2000
          example: "Newly renovated apartment with modern amenities"
        amenities:
          type: array
          items:
            type: string
          example: ["WiFi", "Kitchen", "Parking", "Air Conditioning", "Balcony"]
        isActive:
          type: boolean
          example: true
    PropertyWithReviews:
      type: object
      properties:
        property:
          $ref: '#/components/schemas/Property'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
    PropertySuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Property'
    PropertiesSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Property'
    PropertyWithReviewsSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PropertyWithReviews'

    # Hostaway Types
...